requirements.txt
dune-client==1.7.4
psycopg2-binary>=2.9.3
python-dotenv>=0.20.0
requests>=2.28.1
pandas==2.1.4
ndjson>=0.3.1
py-multiformats-cid>=0.4.4
boto3>=1.26.12
SQLAlchemy>=2.0,<3.0
web3==6.20.3
pandas-stubs>=1.5.1.221024
boto3-stubs>=1.26.12
black==23.12.0
mypy==1.3.0
mypy-extensions==1.0.0
pylint>=2.14.4
pytest>=7.1.2
sqlalchemy-stubs>=0.4
